WIP Moving fragments to a single definition
This work abstracts the differences between inputs
and ouputs in terms of the fragments required to
build them.

This is working to the goal of modeling SSL 
settings in addition to the existing settings
such as the files to index in inputs.conf, and
the head nodes to connect to in ouputs.conf.

The high level defined resource types need to be
implemented at this point, building upon the
consolidated splunk::fragment defined type.